@@ -1,17 +1,22 @@
 package com.kyoapps.maniac.ui.adapters
 
+import android.content.Context
+import android.graphics.Color
 import android.os.Bundle
 import android.os.Handler
-import android.support.v4.widget.SlidingPaneLayout
-import android.support.v7.recyclerview.extensions.ListAdapter
-import android.support.v7.util.DiffUtil
-import android.support.v7.widget.RecyclerView
+import androidx.annotation.ColorInt
+import androidx.slidingpanelayout.widget.SlidingPaneLayout
+import androidx.recyclerview.widget.ListAdapter
+import androidx.recyclerview.widget.DiffUtil
+import androidx.recyclerview.widget.RecyclerView
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
+import android.widget.Toast
 import com.kyoapps.maniac.R
+import com.kyoapps.maniac.dagger.components.DaggerActivityComponent
 
 import com.kyoapps.maniac.functions.FuncFetch
 import com.kyoapps.maniac.functions.FuncUi
@@ -21,36 +26,44 @@ import com.kyoapps.maniac.room.entities.ThreadEnt
 import com.kyoapps.maniac.viewmodel.MainDS
 import com.kyoapps.maniac.viewmodel.MainVM
 
-class MainThreadsAdapter(private val slidingPaneLayout: SlidingPaneLayout?, private val mainVM: MainVM,
-                         private val mainDS: MainDS, private val colorSelected: Int, private val colorPressed: Int,
-                         private val settings: ThreadDisplaySettingsItem)
-    : ListAdapter<ThreadEnt, RecyclerView.ViewHolder>(DIFF_CALLBACK) {
+class MainThreadsAdapter(val context: Context?, private val slidingPaneLayout: androidx.slidingpanelayout.widget.SlidingPaneLayout?, private val component: DaggerActivityComponent)
+    : ListAdapter<ThreadEnt, androidx.recyclerview.widget.RecyclerView.ViewHolder>(DIFF_CALLBACK) {
 
     private var lastSelectedId = -1L
+
+    @ColorInt private val colorSelected = FuncUi.getAttrColorData(context, R.attr.colorPrimary)
+    @ColorInt private val colorNotSelectedUnread = context?.resources?.getColor(R.color.grey_trans_2) ?: Color.GRAY
+    @ColorInt private val textColorNotSelectedUnread = FuncUi.getAttrColorData(context, R.attr.textColorStrong)
+    @ColorInt private val textColorNotSelectedRead = FuncUi.getAttrColorData(context, R.attr.textColorDim)
+    private val textBackgroundUnread = FuncUi.makeColorStateList(Color.WHITE, Color.WHITE, textColorNotSelectedUnread)
+    private val textBackgroundRead = FuncUi.makeColorStateList(Color.WHITE, Color.WHITE, textColorNotSelectedRead)
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ThreadViewHolder {
         val layoutInflater = LayoutInflater.from(parent.context)
         val view = layoutInflater.inflate(R.layout.main_thread_row, parent, false)
-        view.background = FuncUi.makeSelector(colorSelected, colorPressed)
+        view.background = FuncUi.makeSelector(colorSelected, colorSelected, colorNotSelectedUnread)
 
         return ThreadViewHolder(view)
     }
     
-    override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
+    override fun onBindViewHolder(holder: androidx.recyclerview.widget.RecyclerView.ViewHolder, position: Int) {
         val threadEnt = getItem(position)
         if (threadEnt != null) {
             (holder as ThreadViewHolder).bindTo(threadEnt)
+            holder.title.setTextColor(textBackgroundUnread)
+            holder.count.setTextColor(textBackgroundUnread)
             holder.view.isSelected = lastSelectedId == getItemId(position)
         }
 
         (holder as ThreadViewHolder).view.setOnClickListener { _->
             getItem(holder.adapterPosition)?.let {
-                Log.i(TAG, "clicked MainThreadsAdapter brdid ${it.brdid} thrdid ${it.thrdid}")
+                Log.i(TAG, "read MainThreadsAdapter brdid ${it.brdid} thrdid ${it.thrdid}")
 
                 if (it.thrdid.toLong() != lastSelectedId) {
                     setLastSelected(holder, it.thrdid.toLong())
-                    FuncFetch.fetchReplies(mainDS, LoadRequestItem(it.brdid, it.thrdid, null))
-                    mainVM.setRepliesRequestItem(LoadRequestItem(it.brdid, it.thrdid, null))
+                    FuncFetch.fetchReplies(component.mainVM, component.mainDS, LoadRequestItem(it.brdid, it.thrdid, null))
+                    component.mainVM.setRepliesRequestItem(LoadRequestItem(it.brdid, it.thrdid, null))
+                    //Toast.makeText(context, "click; request: ${LoadRequestItem(it.brdid, it.thrdid, null)}", Toast.LENGTH_SHORT).show()
                 }
                 Handler().postDelayed({slidingPaneLayout?.closePane()}, 480)
             }
@@ -58,7 +71,7 @@ class MainThreadsAdapter(private val slidingPaneLayout: SlidingPaneLayout?, priv
     }
 
 
-    class ThreadViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
+    class ThreadViewHolder(itemView: View) : androidx.recyclerview.widget.RecyclerView.ViewHolder(itemView) {
         var title: TextView = itemView.findViewById(R.id.tv_thread_row_title)
         var count: TextView = itemView.findViewById(R.id.tv_thread_row_replycount)
         var view = itemView
@@ -104,11 +117,11 @@ class MainThreadsAdapter(private val slidingPaneLayout: SlidingPaneLayout?, priv
                             && oldThread.oldReplies == newThread.oldReplies
                             && oldThread.totalReplies == newThread.totalReplies
 
-            override fun getChangePayload(oldItem: ThreadEnt?, newItem: ThreadEnt?): Any? {
+            override fun getChangePayload(oldItem: ThreadEnt, newItem: ThreadEnt): Any? {
                 val diffBundle = Bundle()
-                if (newItem?.subject != oldItem?.subject)  diffBundle.putString("KEY_SUBJECT", newItem?.subject)
-                if (newItem?.totalReplies != oldItem?.totalReplies) diffBundle.putInt("KEY_TOTAL_REPLIES", newItem?.totalReplies?: 0)
-                if (newItem?.oldReplies != oldItem?.oldReplies) diffBundle.putInt("KEY_OLD_REPLIES", newItem?.totalReplies?: 0)
+                if (newItem.subject != oldItem.subject)  diffBundle.putString("KEY_SUBJECT", newItem.subject)
+                if (newItem.totalReplies != oldItem.totalReplies) diffBundle.putInt("KEY_TOTAL_REPLIES", newItem.totalReplies)
+                if (newItem.oldReplies != oldItem.oldReplies) diffBundle.putInt("KEY_OLD_REPLIES", newItem.totalReplies)
                 if (diffBundle.size() == 0) return null
                 return diffBundle
             }
