@@ -1,7 +1,7 @@
 package com.kyoapps.maniac.room.entities
 
-import android.arch.persistence.room.Entity
-import android.arch.persistence.room.PrimaryKey
+import androidx.room.Entity
+import androidx.room.PrimaryKey
 
 @Entity(tableName = "reply")
 data class ReplyEnt(
@@ -13,6 +13,7 @@ data class ReplyEnt(
         val subject: String,
         val user: String,
         val replyTime: String,
-        var clicked: Boolean,
+        var read: Boolean,
         val fetchTime: Long
 )
+
