@@ -1,7 +1,7 @@
 package com.kyoapps.maniac.room.entities
 
-import android.arch.persistence.room.Entity
-import android.arch.persistence.room.PrimaryKey
+import androidx.room.Entity
+import androidx.room.PrimaryKey
 
 @Entity(tableName = "thread")
 data class ThreadEnt(
